migrate styles to chakra
