--- web/components/AccessManagement/AccessManagementPage.module.css
@@ -1,44 +1,0 @@
-.buttonRow {
-  display: flex;
-  flex-direction: row;
-}
-
-.unselectedPermission {
-  background-color: #ffffff;
-  color: black;
-}
-
-.selectedPermission {
-  background-color: #a9a8a8;
-  color: white;
-}
-
-.tunselectedPermission {
-  color: black;
-}
-
-.tselectedPermission {
-  color: white;
-}
-
-.selectButtonFilled {
-  background-color: #838282;
-  border: solid;
-  border-color: #ffffff;
-  border-width: 5px;
-}
-
-.selectButtonUnfilled {
-  background-color: "#EEEEEE";
-  border: solid;
-  border-color: #000000;
-  border-width: 1px;
-}
-
-.outlineItem {
-  border: 1.5px solid #000000;
-}
-
-.noOutlineItem {
-  border: none;
-}
